Partially revert 1.1.0, fixing the risk of corrupting the original file while still allowing for files with single quotes and other special characters.
Basically, we couldn't create symbolic links to do this before because Windows requires administrator privileges for that.
Well, it turns out hard links do not require any extra permissions.
It still creates an exra file in the working directory which is unfortunate, but we can't shove it off in the system's temporary files directory because
that could be on another drive and hard links have to be in the same filesystem.